feat(compiler): contract-pass plan — Phase 0 + Workstreams A–F + paged-KV native rung
Close the recurring meta-gap (typed contracts with no consuming pass) across the
8-item KV/attention/quant/TP/multimodal audit. Each contract now has a consuming
pass AND a semantics-preserving oracle, tracked by a drift-gated dashboard.

Phase 0 — compiler/contract_consumers.py: live-probed meta-gap dashboard
  (docs/audit/generated/contract_consumers.{md,csv}); reports 6 live / 0 declared.
A — cache/paged_kv.py: PagedKVState ABI unifying contiguous + tiered KV;
  ops.paged_attention + flash_attn(kv_state=); evaluator.paged_kv_equivalence.
A follow-on (#8) — native Apple-GPU (Metal 4.0) paged attention via the shipped
  fused matmul→softmax→matmul kernel with provenance gating; LATENT (MLA expand)
  + QUANTIZED_TAIL kinds; evaluator.paged_kv_native_equivalence (native rung).
B — compiler/phase_specialization.py: prefill/decode as distinct schedules +
  CacheHandoff ABI; @jit(phase=,slo=); verify_phase_split oracle.
C — fusion.select_attention_lowering: IO-byte selector replaces the hard
  Nk<=SYNTH_MAX_N branch; paged_stage_bytes feeds page-gather cost.
D — compiler/smoothquant.py: W8A8 activation-scale migration producer pass +
  anti-fallback oracle (operands stay int8).
E — compiler/tensor_parallel.py: column/row/sequence-parallel rewrite_linear +
  cross-rank gradient-equivalence oracle (MockRankGroup).
F — compiler/model_walk.py: named multimodal walks (partition_walks) + first-class
  encoder-free ops (patch/coordinate/audio projection) + walk-parity oracle.

Also: numpy-2 uint8-safety fix in quantization.py int4/fp4 packers (pre-existing
working-tree change, ridden along).

Tests: ~120 new unit tests across the new modules; mypy clean; 17 generated docs
in sync (drift-gated).

+++ python/tessera/quantization.py
@@ -189,7 +189,7 @@ def quantize_int4_packed(w: Any, group_size: int = 64):
     evens = codes[:, 0:k:2]                  # length ceil(K/2) = pitch
     odds = codes[:, 1:k:2]                    # length floor(K/2)
     packed = evens.astype(np.uint8, copy=True)
-    packed[:, : odds.shape[1]] |= (odds.astype(np.uint8) << 4)
+    packed[:, : odds.shape[1]] |= (odds.astype(np.uint8) << 4).astype(np.uint8)
     return packed, scales, biases
 
 
@@ -239,7 +239,7 @@ def quantize_fp4_packed(w: Any, group_size: int = 32, scale_mode: str = "mx"):
     evens = codes[:, 0:k:2]
     odds = codes[:, 1:k:2]
     packed = evens.astype(np.uint8, copy=True)
-    packed[:, : odds.shape[1]] |= (odds.astype(np.uint8) << 4)
+    packed[:, : odds.shape[1]] |= (odds.astype(np.uint8) << 4).astype(np.uint8)
     return packed, scales
